Fix Magic Man's E inaccuracy, change practice bot
